+++ src/styles/landing.css
@@ -1,7 +1,7 @@
 .full-background {
     height: 100vh;
     width: 100vw;
-    background-image:linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url('./background.jpg');
+    background-image:linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url('../assets/img/background/background.jpg');
     background-size: cover;
     background-position: center;
     display: flex;
